Added all timetables, courses filtered by semester
All courses which have available timetables have had their timetables added. There are a few courses which are missing timetables on arion, we can fake this data later.
Courses are now grouped by semester on construction of the courseCollectionManager. This will make it easy to return lists filtered by semester