Initial link to view
Refactored the view and its sub-panels to allow linking of the view to the model.
There were a number of gui elements which were created in the View constructor but their references were not held, this was fixed.

All gui elements now have the controller assigned as an actionlistener

Controller now checks for ActionCommands instead of checking object references. This allows us to keep gui elements in the model.