Integrated clash checking with view
Selecting a stream now prompts the model to check whether this stream will clash with any other of the students enrolment. After checks are made it will notify the view of the result.
View will enable and disable the add to schedule button based on the result of the clash check.