Fixed login popup and connected to model
Original implementation of login window by clayton was impossible to connect to the database as it would force the application to hang until a login was entered. Because of this the database was not reachable from the login window and the login could not be connected to the controller.

I reworked the login popup to implement Observer which enabled connection to the model.
I also moved the login popup inside the view class to allow construction of the database and controller prior to the popup opening.