Fixed bug where remove enrolments showed on both tabs
View constructor was a mess with nonsense names and some panels being constructed twice for no reason.

The panel for removing courses was being constructed twice and then added to both the course adding panel and the removal tab.

I have made all panels final to avoid this in the future.